added type checking for call arguments

+++ quark/test/emit/simple-math/java/Functions.java
@@ -4,11 +4,11 @@ public class Functions {
         System.out.println(Integer.toString(four));
         Integer half = ~((~(1)) / (2));
         System.out.println(Integer.toString(half));
-        Double num = (Double) (314.0);
-        Double den = (Double) (100.0);
+        Double num = 314.0;
+        Double den = 100.0;
         Double pi = (num) / (den);
         System.out.println(Double.toString(pi));
-        Double pie = (Double) (3.14);
+        Double pie = 3.14;
         System.out.println(Double.toString(pie));
         Integer n = ~((~(-(100))) / (3));
         System.out.println(Integer.toString(n));
